refactor some hook names
